Fixed image rendering issue and added styling for button and iframe.
