+++ package.json
@@ -18,6 +18,8 @@
   },
   "homepage": "https://github.com/harshabc4/monday-montivation#readme",
   "dependencies": {
+    "ejs": "^3.1.8",
+    "express": "^4.18.1",
     "figlet": "^1.5.2"
   }
 }
